# openapi/hub.json
{
  "components": {
    "schemas": {
...
      "CatalogEdit": {
        "description": "Exactly one of family and variant; a typed value, or unavailable=true.",
        "properties": {
          "attribute_key": {
            "title": "Attribute Key",
            "type": "string"
          },
          "family_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "unavailable": {
            "default": false,
            "title": "Unavailable",
            "type": "boolean"
          },
          "value": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TypedValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          }
        },
        "required": [
          "attribute_key"
        ],
        "title": "CatalogEdit",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": true,
        "description": "Every refusal has this shape; a 409 adds its `code` and may add ids to act on.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why a 409 happened, e.g. VERSION_CONFLICT or ASSESSMENT_OPEN.",
            "title": "Code"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "FamilyView": {
        "properties": {
          "brand_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Name"
          },
          "category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Code"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "manufacturer": {
            "title": "Manufacturer",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Type"
          },
          "properties_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties Text"
          },
          "source_document": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document"
          },
          "source_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Page"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantView"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "manufacturer",
          "brand_name",
          "product_type",
          "category_code",
          "description",
          "properties_text",
          "source_document",
          "source_page",
          "variants"
        ],
        "title": "FamilyView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OwnFact": {
        "description": "A value the supplier set itself (or its \"not available\"), which it may withdraw.",
        "properties": {
          "attribute_key": {
            "title": "Attribute Key",
            "type": "string"
          },
          "fact_id": {
            "format": "uuid",
            "title": "Fact Id",
            "type": "string"
          },
          "unavailable": {
            "title": "Unavailable",
            "type": "boolean"
          },
          "value": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "variant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant Id"
          }
        },
        "required": [
          "fact_id",
          "attribute_key",
          "variant_id",
          "value",
          "unavailable"
        ],
        "title": "OwnFact",
        "type": "object"
      },
...
      "VariantAttributesView": {
        "properties": {
          "additional_information": {
            "additionalProperties": true,
            "title": "Additional Information",
            "type": "object"
          },
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "identifiers": {
            "items": {},
            "title": "Identifiers",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "variant_id": {
            "format": "uuid",
            "title": "Variant Id",
            "type": "string"
          }
        },
        "required": [
          "variant_id",
          "label",
          "attributes",
          "identifiers",
          "additional_information"
        ],
        "title": "VariantAttributesView",
        "type": "object"
      },
...
      "VariantView": {
        "properties": {
          "article_no": {
            "title": "Article No",
            "type": "string"
          },
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "category_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Code"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "supplier": {
            "title": "Supplier",
            "type": "string"
          },
          "variant_id": {
            "format": "uuid",
            "title": "Variant Id",
            "type": "string"
          }
        },
        "required": [
          "variant_id",
          "article_no",
          "display_name",
          "category_code",
          "supplier",
          "attributes"
        ],
        "title": "VariantView",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Supplier hub",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/families/{family_id}": {
      "get": {
        "operationId": "get_family_api_v1_catalog_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Family",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/variants": {
      "get": {
        "operationId": "list_variants_api_v1_catalog_variants_get",
        "parameters": [
          {
            "in": "query",
            "name": "supplier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VariantView"
                  },
                  "title": "Response List Variants Api V1 Catalog Variants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Variants",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/variants/{variant_id}/attributes": {
      "get": {
        "description": "What the node's current-product preview is built from: values with their fact ids (\u00a78.2).",
        "operationId": "variant_attributes_api_v1_catalog_variants__variant_id__attributes_get",
        "parameters": [
          {
            "in": "path",
            "name": "variant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Variant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantAttributesView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Variant Attributes",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/supplier/catalog": {
      "get": {
        "description": "A supplier sees its own catalog and nothing else.",
        "operationId": "own_catalog_api_v1_supplier_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FamilyView"
                  },
                  "title": "Response Own Catalog Api V1 Supplier Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Own Catalog",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/supplier/catalog/facts": {
      "put": {
        "description": "Sets a value for the whole family, or overrides it for one variant.",
        "operationId": "set_catalog_value_api_v1_supplier_catalog_facts_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnFact"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Catalog Value",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/supplier/catalog/facts/{fact_id}": {
      "delete": {
        "description": "Takes back one of the supplier's own values; what was there before shows again.",
        "operationId": "withdraw_catalog_value_api_v1_supplier_catalog_facts__fact_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "fact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Fact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Withdraw Catalog Value",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/supplier/catalog/families/{family_id}": {
      "get": {
        "description": "The supplier's family: its own values, and each variant's with its scope (\u00a79).",
        "operationId": "supplier_family_api_v1_supplier_catalog_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierFamilyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Supplier Family",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/suppliers": {
      "get": {
        "operationId": "list_suppliers_api_v1_suppliers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SupplierView"
                  },
                  "title": "Response List Suppliers Api V1 Suppliers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Missing, expired or revoked token"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The caller's role may not do this"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not found, or another tenant's"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A state or version conflict; see `code`"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Suppliers",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}